Simplify layout: flatten agent.py into hosted/, merge bridge into one file
Move src/agent.py -> hosted/responses/agent.py (no other consumer, no more
sys.path hack) and merge the 3-file backend/ bridge into one bridge_app.py.
Updated all skill docs + verify.sh to match; re-verified end-to-end against
a real local hosted agent (12/12 smoke checks, verify.sh all pass).
