+++ skills/foundry-hosted-agent-copilotkit/references/snippets/hosted/responses/main.py
@@ -1,8 +1,8 @@
 # Copyright (c) Microsoft. All rights reserved.
 """Foundry hosted-agent entry point.
 
-Wraps the shared build_hosted_agent() (src/agent.py, the single brain) in a
-ResponsesHostServer so it can run:
+Wraps build_hosted_agent() (agent.py, right next to this file — the single
+brain) in a ResponsesHostServer so it can run:
   - locally, via `azd ai agent run` (this file executed directly), and
   - deployed, as the published Azure AI Foundry hosted agent (same image/code).
 
@@ -13,21 +13,9 @@ illustrate the shape, but the exact package/module name has moved before and
 may move again; trust what the generated scaffold's own `main.py` imports.
 """
 
-import os
-import sys
-
-# Make the project-root `src/` package importable regardless of the current
-# working directory this script is launched from (local `azd ai agent run`
-# runs it with cwd=hosted/responses; a container build also typically
-# preserves this relative layout).
-_ROOT = os.path.abspath(os.path.join(os.path.dirname(__file__), "..", ".."))
-if _ROOT not in sys.path:
-    sys.path.insert(0, _ROOT)
-
-from agent_framework_foundry_hosting import ResponsesHostServer  # noqa: E402
-from dotenv import load_dotenv  # noqa: E402
-
-from src.agent import build_hosted_agent  # noqa: E402
+from agent import build_hosted_agent
+from agent_framework_foundry_hosting import ResponsesHostServer
+from dotenv import load_dotenv
 
 load_dotenv()
