+++ skills/foundry-hosted-agent-copilotkit/references/troubleshooting.md
@@ -57,7 +57,7 @@ what you actually have installed before trusting it.
 | --- | --- | --- |
 | `az acr build` fails `toomanyrequests` | Docker Hub base image | use `mcr.microsoft.com/devcontainers/...` base images. |
 | azd deploys the helloworld placeholder | ran `azd provision` only | run `azd up` (provision + deploy). |
-| hosted image missing `src/agent.py` (or other shared code) | build context too narrow | make sure `hosted/azure.yaml`'s build context includes wherever your shared agent code actually lives. |
+| hosted image missing `agent.py` (or other agent code) | build context too narrow — this shouldn't happen if `agent.py` lives inside `hosted/responses/` (the recommended layout); only occurs if you've moved shared agent code outside `hosted/` | keep `agent.py` next to `main.py` inside `hosted/responses/` so the default build context already includes it; if you deliberately keep shared code elsewhere, widen `hosted/azure.yaml`'s build context to reach it. |
 | container crashes at boot with `ModuleNotFoundError: No module named 'mcp'`, surfacing as HTTP 424 `session_not_ready` on invoke | the hosting package imports `mcp` but it isn't declared/pulled in transitively on a remote build | pin `mcp` explicitly in `requirements.txt`. |
 
 ## Bridge (whatever you name it)
@@ -66,7 +66,7 @@ what you actually have installed before trusting it.
 | --- | --- | --- |
 | Approval card vanishes when a turn makes several tool calls | multi-tool AG-UI snapshot translation issue (see AG-UI rendering above) | verify live against your actual CopilotKit version rather than assuming a fixed rule. |
 | HITL approve does nothing / state doesn't change | the bridge resolved/dropped the approval locally instead of forwarding it | the single most load-bearing bridge behavior — verify with a live test: reject must leave state unchanged, approve must change it. |
-| `useAgent().state` stays empty | `state_schema`/`predict_state_config` not passed to the agent, or no tool writes the state key, or your bridge doesn't relay state deltas (this pattern is roadmap through a pure proxy — see patterns-7.md) | set `AGENT_STATE_SCHEMA`/`AGENT_PREDICT_STATE`-equivalent config in `src/agent.py` and write the key from a tool; confirm your bridge actually forwards state events before assuming this pattern works out of the box. |
+| `useAgent().state` stays empty | `state_schema`/`predict_state_config` not passed to the agent, or no tool writes the state key, or your bridge doesn't relay state deltas (this pattern is roadmap through a pure proxy — see patterns-7.md) | set `AGENT_STATE_SCHEMA`/`AGENT_PREDICT_STATE`-equivalent config in `hosted/responses/agent.py` and write the key from a tool; confirm your bridge actually forwards state events before assuming this pattern works out of the box. |
 | Deployed bridge can't reach the agent | required settings (e.g. the Foundry project endpoint, the hosted agent's name) unset | set whatever your bridge's platform-mode code needs; it should be able to reach the deployed agent keyless via Entra. |
 | Python `@tool` didn't run "in Foundry" | the native `FoundryAgent` client (not the hosted-agent-first `build_hosted_agent` path this skill uses) runs Python `@tool` callables CLIENT-SIDE; only Foundry-native tools run server-side | expected for that client — this skill's `build_hosted_agent`/`FoundryChatClient` path runs `@tool`s server-side in the hosted runtime instead. |
 | UI 500 mid-run on a long silent tool | a gateway dropped the idle SSE connection | add an SSE keep-alive (a periodic `: ping` comment, e.g. every ~10s) to your bridge's streaming response. |
